- Fix warning.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@21565 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/readers/seqdb/seqdbatlas.cpp
@@ -1138,7 +1138,7 @@ void CSeqDBAtlas::UnregisterExternal(CSeqDBMemReg & memreg)
     size_t bytes = memreg.m_Bytes;
     
     if (bytes > 0) {
-        _ASSERT(m_CurAlloc >= bytes);
+        _ASSERT((int)m_CurAlloc >= (int)bytes);
         m_CurAlloc     -= bytes;
         memreg.m_Bytes = 0;
     }
